\name{ElecForecast-package}
\alias{ElecForecast-package}
\alias{ElecForecast}
\docType{package}
\title{
  Electricity consumption forecasting
}
\description{
  Time-series forecasting models focused on electricity consumption
}
\details{
  We focus on two types of models, a Ridge Regression model and a Bayesian Gaussian Process Regression
}
\author{
Kieran, Cecina, Sherman

Maintainer: Sherman Khoo <sherman.khoo@bristol.ac.uk>
}
\references{
Ref.
}
\keyword{ Time-series }
\seealso{
  Optional links to other man pages
}
\examples{
  \dontrun{
     ## Optional simple examples of the most important functions
     ## These can be in \dontrun{} and \donttest{} blocks.   
  }
}
