% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{GenFT}
\alias{GenFT}
\title{Generate Fourier Transform Matrix}
\usage{
GenFT(time_counter, K, period)
}
\arguments{
\item{time_counter}{A numeric vector of time points at which the Fourier terms are evaluated.}

\item{K}{An integer specifying the number of harmonics to include in the transformation.}

\item{period}{A double indicating the period with respect to which the frequencies are calculated.}
}
\value{
A numeric matrix where each row corresponds to a time point in `time_counter`.
        Each row contains `2 * K` entries: for each k from 1 to K, there is a sine term
        followed by a cosine term corresponding to the k-th harmonic.
}
\description{
This function computes the Fourier transform terms for a given vector of time points.
It constructs a matrix with sine and cosine terms up to the specified harmonic order `K`.
Each row corresponds to a time point, with interleaved sine and cosine values for
frequencies from `1` to `K`.
}
\examples{
time_vec <- seq(0, 10, length.out = 100)
K_value <- 5
period_value <- 10
fourier_matrix <- GenFT(time_vec, K_value, period_value)
dim(fourier_matrix) # should be 100 x 10

}
