% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{predict_parallel_ridge_cv}
\alias{predict_parallel_ridge_cv}
\title{Make Predictions Using a Trained Parallel Ridge Regression Model}
\usage{
predict_parallel_ridge_cv(
  model,
  x_test,
  y_test,
  time_counter,
  daily_period,
  annual_period
)
}
\arguments{
\item{model}{A list containing the trained ridge regression model and its parameters.}

\item{x_test}{Numeric matrix of new test predictor variables.}

\item{y_test}{Numeric vector of new test response variables.}

\item{time_counter}{Numeric vector representing time points for generating Fourier terms.}

\item{daily_period}{Double specifying the period for daily Fourier terms.}

\item{annual_period}{Double specifying the period for annual Fourier terms.}
}
\value{
A list containing the mean squared error (`error`) and the predictions.
}
\description{
This function uses a trained parallel ridge regression model to make predictions on new test data. It also computes and returns the mean squared error of the predictions.
}
\examples{
\dontrun{
# Assuming `results` is obtained from `parallel_ridge_cross_validation`
test_data <- matrix(rnorm(200), ncol = 10)
test_response <- rnorm(20)
prediction_results <- predict_parallel_ridge_cv(results$final_model, test_data, test_response, seq(1, 20, by = 1), 1, 365)
print(prediction_results$error)
print(prediction_results$predictions)
}
}
