% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rbf_kernel}
\alias{rbf_kernel}
\title{Radial Basis Function (RBF) Kernel Matrix Computation}
\usage{
rbf_kernel(X, l, sigma)
}
\arguments{
\item{X}{A numeric matrix where each row represents an observation and each column represents a feature.}

\item{l}{A double specifying the length scale of the kernel.}

\item{sigma}{A double specifying the scale parameter of the kernel.}
}
\value{
A symmetric numeric matrix of size equal to the number of rows in `X`.
        This matrix is the RBF kernel matrix, where each element represents the
        similarity between the corresponding pair of observations in `X`.
}
\description{
This function computes the RBF kernel matrix for a given matrix of input features `X`.
The RBF kernel is also known as the Gaussian kernel. It is a measure of similarity
between rows in `X` that decays exponentially with the squared Euclidean distance
between rows. Each element `K(i, j)` of the output matrix `K` is computed as:
\deqn{K(i, j) = \sigma^2 * exp(-||X[i, ] - X[j, ]||^2 / (2 * l^2))}
where `||X[i, ] - X[j, ]||` is the Euclidean distance between row `i` and `j` of `X`,
`l` is the length scale of the kernel, and `sigma` is the scale parameter.
}
\examples{
# Generate some data
X <- matrix(rnorm(50), nrow=10)
# Compute the RBF kernel with length scale 1 and sigma 1
K <- rbf_kernel(X, 1, 1)
print(K)

}
